Backport theme changes by Z.
He tested the theme on the Nightride.FM page as he implemented the theme
engine there. It didn't quite work as I expected and a lot of the changes
seem to just be setting !important flags or fixing paths, but he also
fixed a padding issue with EQs on mobile and some ugliness with
pre-formatted links and strongly-highlighted ("bold") content.

I don't know why he decided to remove the limitation of applying theme
properties just to the main and archives page. That was intentional but
maybe that was also caused by theme engine breakage. Either way, I
removed it here as well.